refactor SearchFood from AddFoodModal to use where else
